[update] v0.0.5: update upstream base from apache/poi@6a8994e (5.2.5 era) to REL_5_5_1 (POI 5.5.1, Nov 2025)
- Regenerate 238 carried sources from apache/poi REL_5_5_1 (094968cf) with fork conventions preserved (m.co.rh.id.apoi_spreadsheet prefix, android.util.Log, no TempFile/awt); carry new Android-safe classes (HSSFParser/XSSFParser, HSSFReadException/XSSFReadException, ReferenceRelationship, OPCComplianceFlags, CleanerUtil, Reproducibility, etc.)

- Android adaptation fixes so poi and poi-ooxml compile on 5.5.1 base

- Dependencies to upstream 5.5.1 baseline: poi-ooxml-full 5.5.1, commons-io 2.21.0, commons-collections4 4.5.0, commons-compress 1.28.0 (with ZipFile.builder() migration), commons-codec 1.20.0, xmlsec 3.0.6; module versions 0.0.4 -> 0.0.5

- Legal/docs: NOTICE/LICENSE years, fork point reference, README changelog; untrack __pycache__

- Instrumented tests: 39 upstream-changed test files ported to JUnit 4 style (32 material changes, 7 already in sync) + 6 new tests (HSSFParser, XSSFParser, HSSFRowCopyRowFrom, WorkdayFunc, CellUtil, OutOfOrderColumns) + new test-data binaries (bug63189/69769/69812.xlsx, duplicate-filename*.xlsx, clusterfuzz testcase)

- Review fixes applied: LICENSE UTF-8 mojibake (byte-parity with master), TestXSSFBugs bug47504 indentation, changelog wording

+++ poi/src/main/java/m/co/rh/id/apoi_spreadsheet/org/apache/poi/util/HexDump.java
@@ -15,6 +15,8 @@
    limitations under the License.
 ==================================================================== */
 
+// Derived from Apache POI (https://github.com/apache/poi @ commit 094968cfc3d48224db08f0b7f0a6fc341b035114); this file has been modified for Android compatibility by the a-poi-spreadsheet project.
+
 package m.co.rh.id.apoi_spreadsheet.org.apache.poi.util;
 
 import java.io.IOException;
@@ -22,6 +24,7 @@ import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
+
 
 /**
  * dump data in hexadecimal format
@@ -294,7 +297,7 @@ public final class HexDump {
         char[] buf = new char[nDigits];
         long acc = value;
         for(int i=nDigits-1; i>=0; i--) {
-            int digit = Math.toIntExact(acc & 0x0F);
+            int digit = (int) (acc & 0x0F);
             buf[i] = (char) (digit < 10 ? ('0' + digit) : ('A' + digit - 10));
             acc >>>= 4;
         }
